hivesim: move options to separate file
